# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SAM3 Auto Dataset Factory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, completed, failed, cancelled, paused]
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
  /jobs/auto-label:
    post:
      operationId: createAutoLabelJob
      tags: [jobs]
      summary: Create an auto-label job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoLabelJobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/resume:
    post:
      operationId: resumeJob
      tags: [jobs]
      summary: Resume a paused job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /exports:
    post:
      operationId: createExport
      tags: [exports]
      summary: Create YOLO export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "201":
          description: Export created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
    get:
      operationId: listExports
      tags: [exports]
      summary: List exports
      parameters:
        - name: datasetId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Export list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Export"
  /exports/{exportId}:
    get:
      operationId: getExport
      tags: [exports]
      summary: Get export by ID
      parameters:
        - name: exportId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Export detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
...
components:
  schemas:
...
    JobClass:
      type: object
      properties:
        class_id:
          type: integer
        class_name:
          type: string
        prompt:
          type: string
      required: [class_id, class_name, prompt]
    Job:
      type: object
      properties:
        id:
          type: string
        dataset_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled, paused]
        progress:
          type: number
        processed_images:
          type: integer
        total_images:
          type: integer
        failed_images:
          type: integer
        classes:
          type: array
          items:
            $ref: "#/components/schemas/JobClass"
        config:
          $ref: "#/components/schemas/AutoLabelConfig"
        error_message:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        started_at:
          type: ["string", "null"]
          format: date-time
        completed_at:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - dataset_id
        - status
        - progress
        - processed_images
        - total_images
        - failed_images
        - classes
        - config
        - created_at
    AutoLabelConfig:
      type: object
      properties:
        output_mode:
          type: string
          enum: [bbox_only, segmentation_only, bbox_and_segmentation]
        confidence_threshold:
          type: number
          default: 0.35
        min_mask_area:
          type: integer
          default: 80
        max_detections_per_image:
          type: integer
          default: 50
        polygon_simplification_epsilon:
          type: number
          default: 0.002
        save_intermediate_masks:
          type: boolean
          default: true
        review_required:
          type: boolean
          default: true
      required: [output_mode]
    AutoLabelJobInput:
      type: object
      properties:
        dataset_id:
          type: string
        classes:
          type: array
          items:
            $ref: "#/components/schemas/JobClass"
        config:
          $ref: "#/components/schemas/AutoLabelConfig"
      required: [dataset_id, classes, config]
    JobList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
...
    Export:
      type: object
      properties:
        id:
          type: string
        dataset_id:
          type: string
        format:
          type: string
          enum: [detection, segmentation, both]
        status:
          type: string
          enum: [pending, running, completed, failed]
        download_url:
          type: ["string", "null"]
        file_size_bytes:
          type: ["integer", "null"]
        image_count:
          type: integer
        annotation_count:
          type: integer
        min_confidence:
          type: number
        include_rejected:
          type: boolean
        created_at:
          type: string
          format: date-time
        completed_at:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - dataset_id
        - format
        - status
        - image_count
        - annotation_count
        - min_confidence
        - include_rejected
        - created_at
    ExportInput:
      type: object
      properties:
        dataset_id:
          type: string
        format:
          type: string
          enum: [detection, segmentation, both]
          default: both
        min_confidence:
          type: number
          default: 0.0
        include_rejected:
          type: boolean
          default: false
        splits:
          type: array
          items:
            type: string
            enum: [train, val, test]
      required: [dataset_id, format]
...